Swapped stdint.h (unavailable on Solaris9) with inttypes.h
git-svn-id: http://svn.osgeo.org/postgis/trunk@2303 b70326c6-7e19-0410-871a-916f4a2858ee

+++ loader/wkb.h
@@ -1,3 +1,5 @@
+#include <inttypes.h>
+
 /*
  * OGC WellKnownBinaryGeometry
  */
@@ -8,7 +10,7 @@
 #define WKB3DOFFSET 0x80000000
 
 typedef unsigned char byte;
-typedef unsigned long int uint32;
+typedef uint32_t uint32;
 
 typedef struct Point_t {
 	double x;
